Removed test cases for plugins from other repo than ui.
