+++ tests/manual/tickets/198/1.html
@@ -1,19 +1,14 @@
-<h2>Contextual Toolbar:</h2>
-<p><button id="button-ct">Start external changes</button></p>
-<div id="editor-ct">
+<h2>External changes (insert)</h2>
+
+<p><button id="button-insert">Start external changes</button></p>
+<div id="editor-insert">
 	<p>EXTERNAL CHANGES WILL START HERE -> This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
 </div>
 
-<h2>Link:</h2>
-<p><button id="button-link">Start external changes</button></p>
-<div id="editor-link">
-	<p>EXTERNAL CHANGES WILL START HERE -> This specification <a href="http://ckeditor.com">defines the 5th</a> major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-</div>
-
-<h2>Image toolbar:</h2>
-<p><button id="button-image-toolbar">Start external changes</button></p>
-<div id="editor-image-toolbar">
-	<p>EXTERNAL CHANGES WILL START HERE -> This specification defines the 5th major revision of the core language</p>
-	<figure class="image"><img src="sample.jpg" alt="CKEditor logo" /></figure>
-	<p>of the World Wide Web: the Hypertext Markup Language (HTML).</p>
+<h2>External changes (delete)</h2>
+<p><button id="button-delete">Start external changes</button></p>
+<div id="editor-delete">
+	<p>This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
+	<p>THIS PARAGRAPH WILL BE REMOVED This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
+	<p>This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
 </div>
