+++ tests/manual/tickets/198/1.md
@@ -1,33 +1,14 @@
 # Issue [#198](https://github.com/ckeditor/ckeditor5-ui/issues/198) manual test.
 
-## Position of balloons should be updated on external changes.
+## Position of the `ContextualToolbar` should be updated on external changes.
 
-### Contextual toolbar
+### Insert:
 
 1. Click `Start external changes` and quickly select some text after the place where external changes will star
 (contextual toolbar should appear).
 2. Observe if balloon is attached to the target selection.
 
-### Link #1
+### Delete:
 
-1. Click `Start external changes` and quickly click on link element to display Link balloon.
-2. Observe if balloon is attached to the target element.
-
-### Link #2
-
-1. Refresh test page.
-2. Click `Start external changes` and quickly select some text after the place where external changes will start and press
-toolbar link button (or `Cmd` + `K`) to display Link balloon.
-3. Observe if balloon is attached to the target selection.
-
-### Image toolbar #1
-
-1. Click `Start external changes` and quickly click on image element to display image toolbar.
-2. Observe if balloon is attached to the target element.
-
-### Image toolbar #2
-
-1. Refresh test page.
-2. Click `Start external changes` and quickly click on image element then choose alt component from toolbar
-to display Text alternative balloon.
-3. Observe if balloon is attached to the target element.
+1. Click `Start external changes` and quickly select some text in the second paragraph (contextual toolbar should appear).
+2. Check if balloon hide and there is no errors on the browser console.
