--- tests/manual/tickets/198/2.html
@@ -1,25 +1,0 @@
-<h2>Contextual Toolbar:</h2>
-<p><button id="button-ct">Start external changes</button></p>
-<div id="editor-ct">
-	<p>This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-	<p>THIS PARAGRAPH WILL BE REMOVE This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-	<p>This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-</div>
-
-<h2>Link:</h2>
-<p><button id="button-link">Start external changes</button></p>
-<div id="editor-link">
-	<p>This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-
-	<p>THIS PARAGRAPH WILL BE REMOVE This specification <a href="http://ckeditor.com">defines the 5th</a> major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-
-	<p>This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-</div>
-
-<h2>Image toolbar:</h2>
-<p><button id="button-image-toolbar">Start external changes</button></p>
-<div id="editor-image-toolbar">
-	<p>This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-	<figure class="image"><img src="sample-to-delete.jpg" alt="CKEditor logo" /></figure>
-	<p>This specification defines the 5th major revision of the core language of the World Wide Web: the Hypertext Markup Language (HTML).</p>
-</div>
