--- tests/manual/tickets/198/2.md
@@ -1,31 +1,0 @@
-# Issue [#198](https://github.com/ckeditor/ckeditor5-ui/issues/198) manual test.
-
-## Balloon should hide on external changes.
-
-### Contextual toolbar
-
-1. Click `Start external changes` and quickly select some text from second paragraph  (contextual toolbar should appear).
-2. Check if balloon hide and there is no errors on browser console.
-
-### Link #1
-
-1. Click `Start external changes` and quickly click on link element from second paragraph  to display Link balloon.
-2. Check if balloon hide and there is no errors on browser console.
-
-### Link #2
-
-1. Refresh test page.
-2. Click `Start external changes` and quickly select some text from second paragraph and press toolbar link button (or `Cmd` + `K`) to display Link balloon.
-2. Check if balloon hide and there is no errors on browser console.
-
-### Image toolbar #1
-
-1. Click `Start external changes` and quickly click on image element to display image toolbar.
-2. Check if balloon hide and there is no errors on browser console.
-
-### Image toolbar #2
-
-1. Refresh test page.
-2. Click `Start external changes` and quickly click on image element then choose alt component from toolbar
-to display Text alternative balloon.
-2. Check if balloon hide and there is no errors on browser console.
